+++ src/main/java/edu/upc/eetac/dsa/music4you/dao/AnuncioDAO.java
@@ -11,7 +11,7 @@ import java.sql.SQLException;
  */
 public interface AnuncioDAO {
 
-    Anuncio createAnuncio(String userid, String subject, String description, long precio, int type, InputStream image) throws SQLException;
+    Anuncio createAnuncio(String userid, String subject, String description, double precio, int type, InputStream image) throws SQLException;
     Anuncio getAnuncioById(String id) throws SQLException;
     AnuncioCollection getAnuncios() throws SQLException;
     Anuncio updateAnuncio(String id, String subject, String description, long precio) throws SQLException;
